+++ packages/cli/openapi-parser/src/openapi/v2/generateIr.ts
@@ -7,16 +7,19 @@ import { generateIr as generateIrFromV3 } from "../v3/generateIr";
 export async function generateIr({
     openApi,
     taskContext,
-    disableExamples
+    disableExamples,
+    audiences
 }: {
     openApi: OpenAPIV2.Document;
     taskContext: TaskContext;
     disableExamples: boolean | undefined;
+    audiences: string[];
 }): Promise<OpenApiIntermediateRepresentation> {
     const conversionResult = await convertObj(openApi, {});
     return generateIrFromV3({
         openApi: conversionResult.openapi,
         taskContext,
-        disableExamples
+        disableExamples,
+        audiences
     });
 }