+++ seed/fastapi/extra-properties/resources/user/service/create_user_request.py
@@ -8,6 +8,8 @@ from ....core.pydantic_utilities import pydantic_v1
 
 
 class CreateUserRequest(pydantic_v1.BaseModel):
+    type: typing.Literal["CreateUserRequest"] = pydantic_v1.Field(alias="_type")
+    version: typing.Literal["v1"] = pydantic_v1.Field(alias="_version")
     name: str
 
     def json(self, **kwargs: typing.Any) -> str:
@@ -19,5 +21,7 @@ class CreateUserRequest(pydantic_v1.BaseModel):
         return super().dict(**kwargs_with_defaults)
 
     class Config:
+        allow_population_by_field_name = True
+        populate_by_name = True
         extra = pydantic_v1.Extra.forbid
         json_encoders = {dt.datetime: serialize_datetime}